handle text files containing a Unicode BOM (byte-order mark).
The Unicode BOM is common under Windows, but rare under Linux
or MacOSX.  Previously the definition file parser (etc) would
fatal error if a BOM was present.  Now it is simply ignored.

+++ src/ui_editor.cc
@@ -824,7 +824,7 @@ bool UI_TextEditor::ContainsUnicode() const
 	int len = tbuf->length();
 
 	for (int i = 0 ; i < len ; i++)
-		if (tbuf->byte_at(i) & 0x80)
+		if ((byte)tbuf->byte_at(i) & 0x80)
 			return true;
 
 	return false;
